requirements.txt
Flask==2.3.3
google-generativeai==0.3.0
gunicorn==21.2.0
Flask-HTTPAuth==4.8.0